chore(parser): move get type and type interface to openapi
