+++ src/openApi/common/interfaces/Type.ts
@@ -1,0 +1,8 @@
+export interface Type {
+    $refs: string[];
+    base: string;
+    imports: string[];
+    isNullable: boolean;
+    template: string | null;
+    type: string;
+}
