+++ src/openApi/v2/parser/getOperationResponse.ts
@@ -1,7 +1,7 @@
 import type { OperationResponse } from '../../../types/client';
 import { getPattern } from '../../../utils/getPattern';
-import { getType } from '../../../utils/type';
 import { getRef } from '../../common/parser/getRef';
+import { getType } from '../../common/parser/type';
 import type { OpenApi } from '../interfaces/OpenApi';
 import type { OpenApiResponse } from '../interfaces/OpenApiResponse';
 import type { OpenApiSchema } from '../interfaces/OpenApiSchema';
