+++ src/openApi/v3/parser/getOperationParameter.ts
@@ -1,8 +1,8 @@
 import type { OperationParameter } from '../../../types/client';
 import { getPattern } from '../../../utils/getPattern';
-import { getType } from '../../../utils/type';
 import { getRef } from '../../common/parser/getRef';
 import { getOperationParameterName } from '../../common/parser/operation';
+import { getType } from '../../common/parser/type';
 import type { OpenApi } from '../interfaces/OpenApi';
 import type { OpenApiParameter } from '../interfaces/OpenApiParameter';
 import type { OpenApiSchema } from '../interfaces/OpenApiSchema';
