+++ src/openApi/v3/parser/getOperationRequestBody.ts
@@ -1,6 +1,6 @@
 import type { OperationParameter } from '../../../types/client';
 import { getPattern } from '../../../utils/getPattern';
-import { getType } from '../../../utils/type';
+import { getType } from '../../common/parser/type';
 import type { OpenApi } from '../interfaces/OpenApi';
 import type { OpenApiRequestBody } from '../interfaces/OpenApiRequestBody';
 import { getContent } from './getContent';
